# Improve performance and tidiness