--- a/src/common.h
+++ b/src/common.h
@@ -92,7 +92,7 @@ extern Score scorer_score(const CandidateString &string, const Element &query, c
 extern Score scoreSize(Score n, Score m);
 
 extern Score path_scorer_score(const CandidateString &string, const Element &query, const Options &options);
-extern int countDir(const CandidateString &path, int end, char pathSeparator);
+extern int countDir(const CandidateString &path, const size_t end, const char pathSeparator);
 extern CandidateString getExtension(const CandidateString &str);
 
 extern const std::vector<CandidateIndex> filter(const vector<std::vector<CandidateString>> &candidates, const Element &query, const Options &options);