--- a/src/fuzzaldrin.cc
+++ b/src/fuzzaldrin.cc
@@ -47,7 +47,7 @@ Napi::Value Fuzzaldrin::SetCandidates(const Napi::CallbackInfo &info) {
     return Napi::Boolean();
 }
 
-void Fuzzaldrin::SetCandidates(vector<CandidateObject> const &candidates) {
+void Fuzzaldrin::SetCandidates(const vector<CandidateObject> &candidates) {
     const auto N = candidates.size();// different
     const auto num_chunks = N < 1000 * kMaxThreads ? N / 1000 + 1 : kMaxThreads;
     candidates_.clear();
@@ -83,16 +83,16 @@ Napi::Value Fuzzaldrin::FilterTree(const Napi::CallbackInfo &info) {
     string const dataKey = info[2].As<Napi::String>();
     string const childrenKey = info[3].As<Napi::String>();
 
-    size_t maxResults = info[4].As<Napi::Number>().Uint32Value();
-    bool usePathScoring = info[5].As<Napi::Boolean>();
-    bool useExtensionBonus = info[6].As<Napi::Boolean>();
+    const size_t maxResults = info[4].As<Napi::Number>().Uint32Value();
+    const bool usePathScoring = info[5].As<Napi::Boolean>();
+    const bool useExtensionBonus = info[6].As<Napi::Boolean>();
 
     // create Tree and set candidates
     auto tree = Tree(jsTreeArray, dataKey, childrenKey);
     SetCandidates(tree.entriesArray);
 
     // create options
-    Options options(query, maxResults, usePathScoring, useExtensionBonus);
+    const Options options(query, maxResults, usePathScoring, useExtensionBonus);
     const auto matches = filter(candidates_, query, options);
 
     // filter