--- a/src/fuzzaldrin.cc
+++ b/src/fuzzaldrin.cc
@@ -60,7 +60,7 @@ void Fuzzaldrin::SetCandidates(const vector<CandidateObject> &candidates) {
             chunk_size++;
         }
         for (auto j = cur_start; j < cur_start + chunk_size; j++) {
-            candidates_[i].push_back(candidates[j].data);// different
+            candidates_[i].push_back(candidates[j].data);// different // TODO copy
         }
         cur_start += chunk_size;
     }
@@ -77,11 +77,11 @@ Napi::Value Fuzzaldrin::FilterTree(const Napi::CallbackInfo &info) {
         Napi::TypeError::New(info.Env(), "Invalid arguments").ThrowAsJavaScriptException();
         return Napi::Array::New(info.Env());
     }
-    auto const jsTreeArray = info[0].As<Napi::Array>();
-    std::string query = info[1].As<Napi::String>();
+    const auto &jsTreeArray = info[0].As<Napi::Array>();
+    const std::string &query = info[1].As<Napi::String>();
 
-    string const dataKey = info[2].As<Napi::String>();
-    string const childrenKey = info[3].As<Napi::String>();
+    const string &dataKey = info[2].As<Napi::String>();
+    const string &childrenKey = info[3].As<Napi::String>();
 
     const size_t maxResults = info[4].As<Napi::Number>().Uint32Value();
     const bool usePathScoring = info[5].As<Napi::Boolean>();
@@ -92,8 +92,8 @@ Napi::Value Fuzzaldrin::FilterTree(const Napi::CallbackInfo &info) {
     SetCandidates(tree.entriesArray);
 
     // create options
-    const Options options(query, maxResults, usePathScoring, useExtensionBonus);
-    const auto matches = filter(candidates_, query, options);
+    const auto options = Options(query, maxResults, usePathScoring, useExtensionBonus);
+    const auto &matches = filter(candidates_, query, options);
 
     // filter
     auto filteredCandidateObjects = Napi::Array::New(info.Env());// array of candidate objects (with their address in index and level)