- Improved Tree struct by not storing the jsTreeObjecrOrArray

This is a continuation of #35 by applying more fixes (mostly manual):
- Avoid copying by using references
- Adds const where possible
- Adds `{}` and `()` to prevent errors
- Add noexcept to some classes
- Use emplace_back instead of push_back
- Adding some TODOs for fixing copying
etc


<details>
<summary>Benchmarks</summary>

TwoLetter#Filter seems faster now.

Benchmarks after:

```
> npm run benchmark:small && npm run benchmark:regular && npm run benchmark:large


> fuzzaldrin-plus-fast@1.1.0 benchmark:small
> node benchmark/benchmark-small.js

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 100 100

====== Running test - query:npm ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 4 ms
length 55 100

====== Running test - query:node ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 100 100

====== Running test - query:grunt ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 33 100

====== Running test - query:html ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 10 100

====== Running test - query:doc ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 5 ms
length 87 100

====== Running test - query:cli ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 2 ms
length 57 100

====== Running test - query:js ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 60 100

====== Running test - query:jas ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 19 100

====== Running test - query:mine ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 3 ms
length 65 100

====== Running test - query:stream ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 19 100


> fuzzaldrin-plus-fast@1.1.0 benchmark:regular
> node benchmark/benchmark.js

====== Running test - query:index ======
Elapsed time - fuzzaldrin-plus-fast: 29 ms vs. fuzzaldrin-plus: 44 ms
length 6168 66672

====== Running test - query:indx ======
Elapsed time - fuzzaldrin-plus-fast: 29 ms vs. fuzzaldrin-plus: 51 ms
length 6192 66672

====== Running test - query:walkdr ======
Elapsed time - fuzzaldrin-plus-fast: 25 ms vs. fuzzaldrin-plus: 16 ms
length 504 66672
====== fuzzaldrin-plus-fast is SLOWER

====== Running test - query:node ======
Elapsed time - fuzzaldrin-plus-fast: 47 ms vs. fuzzaldrin-plus: 69 ms
length 65136 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 45 ms vs. fuzzaldrin-plus: 65 ms
length 65208 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 46 ms vs. fuzzaldrin-plus: 54 ms
length 65208 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 46 ms vs. fuzzaldrin-plus: 59 ms
length 65208 66672

====== Running test - query:ndem ======
Elapsed time - fuzzaldrin-plus-fast: 53 ms vs. fuzzaldrin-plus: 236 ms
length 65124 66672

Matching 66672 results for 'index' took 283ms (Prepare in advance)
Matching 66672 results for 'index' took 280ms (cache)
Matching 66672 results for 'index' took 90ms (legacy)

> fuzzaldrin-plus-fast@1.1.0 benchmark:large
> node benchmark/benchmark-large.js

TwoLetter#legacy: 10.746s
TwoLetter#fuzzaldrin-plus-fast#DirectFilter: 3.209s
TwoLetter#fuzzaldrin-plus-fast#setCandidates#filter: 429.91ms
======
ThreeLetter#legacy: 8.825s
ThreeLetter#fuzzaldrin-plus-fast#DirectFilter: 3.384s
ThreeLetter#fuzzaldrin-plus-fast#setCandidates#filter: 449.717ms
======
TwoLetter#Keybased#Filter: 3.914s
ThreeLetter#Keybased#Filter: 4.579s
======
setCandidates: 224.735ms
TwoLetter#Filter: 422.389ms
ThreeLetter#Filter: 497.014ms
======
setCandidates#Keybased: 229.815ms
TwoLetter#Keybased#Filter: 421.57ms
ThreeLetter#Keybased#Filter: 438.628ms
```

Benchmarks on the previosu released version:

```
> npm run benchmark:small && npm run benchmark:regular && npm run benchmark:large


> fuzzaldrin-plus-fast@1.1.0 benchmark:small
> node benchmark/benchmark-small.js

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 100 100

====== Running test - query:npm ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 4 ms
length 55 100

====== Running test - query:node ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 1 ms
length 100 100

====== Running test - query:grunt ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 33 100

====== Running test - query:html ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 10 100

====== Running test - query:doc ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 4 ms
length 87 100

====== Running test - query:cli ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 2 ms
length 57 100

====== Running test - query:js ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 60 100

====== Running test - query:jas ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 0 ms
length 19 100

====== Running test - query:mine ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 2 ms
length 65 100

====== Running test - query:stream ======
Elapsed time - fuzzaldrin-plus-fast: 0 ms vs. fuzzaldrin-plus: 2 ms
length 19 100


> fuzzaldrin-plus-fast@1.1.0 benchmark:regular
> node benchmark/benchmark.js

====== Running test - query:index ======
Elapsed time - fuzzaldrin-plus-fast: 29 ms vs. fuzzaldrin-plus: 44 ms
length 6168 66672

====== Running test - query:indx ======
Elapsed time - fuzzaldrin-plus-fast: 30 ms vs. fuzzaldrin-plus: 52 ms
length 6192 66672

====== Running test - query:walkdr ======
Elapsed time - fuzzaldrin-plus-fast: 27 ms vs. fuzzaldrin-plus: 16 ms
length 504 66672
====== fuzzaldrin-plus-fast is SLOWER

====== Running test - query:node ======
Elapsed time - fuzzaldrin-plus-fast: 51 ms vs. fuzzaldrin-plus: 68 ms
length 65136 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 51 ms vs. fuzzaldrin-plus: 68 ms
length 65208 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 50 ms vs. fuzzaldrin-plus: 51 ms
length 65208 66672

====== Running test - query:nm ======
Elapsed time - fuzzaldrin-plus-fast: 51 ms vs. fuzzaldrin-plus: 58 ms
length 65208 66672

====== Running test - query:ndem ======
Elapsed time - fuzzaldrin-plus-fast: 51 ms vs. fuzzaldrin-plus: 234 ms
length 65124 66672

Matching 66672 results for 'index' took 289ms (Prepare in advance)
Matching 66672 results for 'index' took 286ms (cache)
Matching 66672 results for 'index' took 84ms (legacy)

> fuzzaldrin-plus-fast@1.1.0 benchmark:large
> node benchmark/benchmark-large.js

TwoLetter#legacy: 10.584s
TwoLetter#fuzzaldrin-plus-fast#DirectFilter: 3.244s
TwoLetter#fuzzaldrin-plus-fast#setCandidates#filter: 483.943ms
======
ThreeLetter#legacy: 8.889s
ThreeLetter#fuzzaldrin-plus-fast#DirectFilter: 3.368s
ThreeLetter#fuzzaldrin-plus-fast#setCandidates#filter: 431.151ms
======
TwoLetter#Keybased#Filter: 3.992s
ThreeLetter#Keybased#Filter: 4.630s
======
setCandidates: 223.59ms
TwoLetter#Filter: 636.285ms
ThreeLetter#Filter: 434.577ms
======
setCandidates#Keybased: 234.892ms
TwoLetter#Keybased#Filter: 498.664ms
ThreeLetter#Keybased#Filter: 444.204ms
```

</details>